+++ src/SmartHopper.Core.Grasshopper/AITools/list_evaluate.cs
@@ -16,7 +16,11 @@ using System.Threading.Tasks;
 using Grasshopper.Kernel.Types;
 using Newtonsoft.Json.Linq;
 using SmartHopper.Core.Grasshopper.Utils;
-using SmartHopper.Infrastructure.AICall;
+using SmartHopper.Infrastructure.AICall.Core.Base;
+using SmartHopper.Infrastructure.AICall.Core.Interactions;
+using SmartHopper.Infrastructure.AICall.Core.Requests;
+using SmartHopper.Infrastructure.AICall.Core.Returns;
+using SmartHopper.Infrastructure.AICall.Tools;
 using SmartHopper.Infrastructure.AIModels;
 using SmartHopper.Infrastructure.AITools;
 
@@ -184,3 +188,4 @@ namespace SmartHopper.Core.Grasshopper.AITools
         }
     }
 }
+